pyproject.toml
[project]
name = "llmwiki"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "langchain-core>=0.3,<0.4",
    "langgraph>=0.3,<0.4",
    "langchain-anthropic>=0.3,<0.4",
    "langchain-google-genai>=2.1,<3",
    "langchain-openai>=0.3,<0.4",
    "pydantic>=2.0,<3",
    "python-dotenv>=1.0,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
]

[project.scripts]
llmwiki = "llmwiki.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llmwiki"]
